Copy dataset definition from research-template
This copies the ehrQL dataset definition from
opensafely/research-template (2694785). The dataset definition is
referenced in the project pipeline by the generate_dataset action, which
replaces the generate_study_population action.

This doesn't remove the cohort-extractor study definition, because we
will update the dataset definition to match the study definition in a
follow-up commit.