Remove study definition
With the exception of the BMI variable, the cohort-extractor study
definition has now been replaced by the ehrQL dataset definition.

The reason for not including the BMI variable is that ehrQL doesn't
include a method for computing BMI values. Instead, BMI values should be
identified using a codelist. Doing that in the dataset definition would
introduce a significant deviation from the study definition.